refactored layout and added loader to view
