+++ src/views/SponsorChild/Sponsor.css
@@ -1,0 +1,3 @@
+.loaderContainer {
+    margin-top: 10%;
+}
